+++ examples/openai/celest/pubspec.yaml
@@ -19,5 +19,5 @@ dependency_overrides:
     path: ../../../packages/celest_core
 
 dev_dependencies:
-  lints: ^3.0.0
-  test: ^1.24.0
+  lints: ^4.0.0
+  test: ^1.25.0
