+++ packages/celest/example/lib/main.dart
@@ -1,23 +1,8 @@
-import 'dart:developer';
-
 import 'package:celest_backend/client.dart';
+import 'package:celest_backend/models/person.dart';
 import 'package:flutter/material.dart';
-import 'package:logging/logging.dart';
 
 void main() {
-  Logger.root.level = Level.ALL;
-  Logger.root.onRecord.listen((record) {
-    log(
-      record.message,
-      time: record.time,
-      sequenceNumber: record.sequenceNumber,
-      level: record.level.value,
-      name: record.loggerName,
-      zone: record.zone,
-      error: record.error,
-      stackTrace: record.stackTrace,
-    );
-  });
   celest.init();
   runApp(const MyApp());
 }
@@ -35,11 +20,11 @@ class MyApp extends StatelessWidget {
         ),
         body: Center(
           child: FutureBuilder(
-            future: celest.functions.greeting.sayHelloMany(
-              people: const ['Alice', 'Bob', 'Charlie'],
-            ).toList(),
+            future: celest.functions.greeting.sayHello(
+              person: const Person(name: 'Celest'),
+            ),
             builder: (_, snapshot) => switch (snapshot) {
-              AsyncSnapshot(:final data?) => Text(data.join('\n')),
+              AsyncSnapshot(:final data?) => Text(data),
               AsyncSnapshot(:final error?) =>
                 Text('${error.runtimeType}: $error'),
               _ => const CircularProgressIndicator(),
