+++ packages/celest/pubspec.yaml
@@ -15,7 +15,7 @@ dependencies:
   chunked_stream: ^1.4.2
   collection: ^1.18.0
   http: ^1.0.0
-  http_parser: ^4.0.2
+  http_parser: ^4.0.0
   logging: ^1.2.0
   meta: ^1.11.0
   shelf: ^1.4.1
