+++ packages/celest_auth/example/celest/pubspec.yaml
@@ -19,4 +19,4 @@ dependency_overrides:
 
 dev_dependencies:
   lints: ^4.0.0
-  test: ^1.24.0
+  test: ^1.25.0
