+++ packages/celest_core/lib/_internal.dart
@@ -4,4 +4,5 @@ export 'src/base/celest_base.dart';
 export 'src/events/sse/sse_client.dart';
 export 'src/http/celest_http_client.dart';
 export 'src/util/globals.dart';
+export 'src/util/json.dart';
 export 'src/util/uuid.dart';
