+++ packages/celest_core/lib/src/http/celest_http_client.dart
@@ -1,6 +1,3 @@
-import 'dart:developer';
-import 'dart:isolate';
-
 import 'package:celest_core/_internal.dart';
 import 'package:celest_core/src/auth/authenticator.dart';
 import 'package:celest_core/src/http/http_client.vm.dart'
@@ -27,30 +24,6 @@ final class CelestHttpClient extends http.BaseClient {
     final cork = await _authenticator.token;
     if (cork != null) {
       request.headers['authorization'] = 'Bearer $cork';
-    }
-    // Support setting breakpoints in Celest functions when debugging the
-    // parent application.
-    //
-    // This is only enabled in debug mode when targeting the local environment.
-    // TODO(dnys1): Use celest.env to make sure we're in local env.
-    if (kDebugMode) {
-      final serviceInfo = await Service.getInfo();
-      final (majorVersion, minorVersion) =
-          (serviceInfo.majorVersion, serviceInfo.minorVersion);
-      request.headers['X-Celest-Debug-Server-Protocol-Major'] = '$majorVersion';
-      request.headers['X-Celest-Debug-Server-Protocol-Minor'] = '$minorVersion';
-      if (serviceInfo.serverUri case final serverUri?) {
-        request.headers['X-Celest-Debug-Server-Uri'] = serverUri.toString();
-      }
-      if (serviceInfo.serverWebSocketUri case final serverWsUri?) {
-        request.headers['X-Celest-Debug-Server-WebSocket-Uri'] =
-            serverWsUri.toString();
-      }
-      if (!kIsWeb) {
-        if (Service.getIsolateId(Isolate.current) case final isolateId?) {
-          request.headers['X-Celest-Debug-Isolate-Id'] = isolateId;
-        }
-      }
     }
     return _inner.send(request);
   }
