+++ packages/celest_core/pubspec.yaml
@@ -12,7 +12,7 @@ dependencies:
   async: ^2.11.0
   collection: ^1.18.0
   http: ^1.0.0
-  http_parser: ^4.0.2
+  http_parser: ^4.0.0
   logging: ^1.2.0
   meta: ^1.10.0
   native_storage: ^0.1.0
@@ -25,4 +25,4 @@ dependencies:
 
 dev_dependencies:
   lints: ^4.0.0
-  test: ^1.24.0
+  test: ^1.25.0
